fix(schema): use Integer for variantSetDbId in storage service
Update GenotypeStorageService and its implementation to use Integer
for variantSetDbId, matching the VariantSet entity. This fixes a
compilation error in AlleleMatrixService introduced by the previous
type change.

+++ src/main/java/org/irri/snpseek/brapi/controller/VariantSetController.java
@@ -176,7 +176,7 @@ public class VariantSetController {
 
     private VariantSetDto toDto(VariantSet vs) {
         String name = vs.getName();
-        long   variantSetDbId = vs.getVariantSetId().longValue();
+        Integer variantSetDbId = vs.getVariantSetId();
 
         // Count SNPs and samples for this variant set
         Specification<SnpMetadata> snpSpec =
